chore: clean up latent compiler warnings (unused vars, unused captures)
- block_wiedemann.cpp: drop unused 'locals' alias, simplify lambda
  captures to access scratch.locals directly. Drop unused
  'nonzero_seqs' diagnostic counter (only trivial_seqs is printed).
- test_perf_targets.cpp: drop unused 'quick'/'siqs_only' flags (their
  effects already encoded in start_digits/end_digits).
- test_integration.cpp: drop redundant cast to int64_t when comparing
  uint64_t fields (rel->b == b — both uint64_t, no cast needed).

All previously-emitted warnings outside implicit-conversion category
now silenced. BW 27/27 + Gate 29/29 PASS.

+++ tests/test_integration.cpp
@@ -826,7 +826,7 @@ void test_sieve_cofactor_joint() {
         if (rel.has_value()) {
             verified++;
             assert(rel->a == a);
-            assert(rel->b == static_cast<int64_t>(b));
+            assert(rel->b == b);  // both uint64_t
         }
     }
